backend/pyproject.toml
[tool.poetry]
name = "django_react_todo"
version = "0.1.0"
description = "A simple TODO Application made with Django & React JS"
authors = ["NumanIbnMazid <numanibnmazid@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
python-dotenv = "^0.19.2"
autopep8 = "^1.6.0"
psycopg2-binary = "^2.9.3"
dj-database-url = "^0.5.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-cors-headers = "^3.10.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
